requirements.txt
GitPython==3.1.41
Flask==3.1.0
Flask-SocketIO==5.5.1
requests==2.31.0
Jinja2==3.1.2
netmiko==4.1.2
PyYAML==6.0
ttp==0.9.5